pull changes

+++ app.py
@@ -217,6 +217,7 @@ def to_do_list():
     else:
         return redirect('/')
 
+"""
 @route('/delete_to_do_list', method=['GET', 'POST'])
 def delete_to_do_list():
     logged_in_cookie = request.get_cookie('loggedIn')
@@ -240,7 +241,7 @@ def delete_to_do_list():
             db.close()
     else:
         return redirect('/')
-
+"""
 @route('/create_to_do_list', method='POST')
 def create_to_do_list():
     logged_in_cookie = request.get_cookie('loggedIn')
@@ -264,8 +265,8 @@ def create_to_do_list():
     else:
         return redirect('/')
 
-@delete('/delete_to_do_list/<to_do_list_id:int>', method="DELETE")
-def delete_to_do_list(to_do_list_id):
+@delete('/delete_to_do_list', method="DELETE")
+def delete_to_do_list():
     logged_in_cookie = request.get_cookie('loggedIn')
     if logged_in_cookie:
         try:
@@ -273,10 +274,16 @@ def delete_to_do_list(to_do_list_id):
             db = make_db_connection()
             cursor = db.cursor()
 
+            to_do_list_id = request.forms.get("to_do_list_id")
+
+
             cursor.execute('DELETE FROM to_do_lists_task WHERE category_id = %s;', (to_do_list_id,))
             db.commit()
-            cursor.execute('DELETE FROM to_do_list WHERE id = %s', (to_do_list_id,))
+            cursor.execute('DELETE FROM to_do_list WHERE id = %s  AND user_id = %s', (to_do_list_id, logged_in_cookie))
             db.commit()
+
+            print("todolist id:", to_do_list_id)
+
             return template('/to_do_list')
         
         finally:
@@ -339,7 +346,7 @@ def update_task_status():
 
     else:
         return redirect('/')
-         
+
 @route('/update_checkboxes')
 def update_checkboxes():
     logged_in_cookie = request.get_cookie('loggedIn')
@@ -359,6 +366,8 @@ def update_checkboxes():
     else:
         return redirect('/')
             
+# Calendar Start
+
 @route('/calendar')
 def calendar():
     logged_in_cookie = request.get_cookie('loggedIn')
@@ -375,6 +384,7 @@ def calendar():
             db.close()
     else:
         return redirect('/')
+        
     
 @route('/add_event', method=['GET', 'POST'])
 def add_event():
@@ -385,52 +395,85 @@ def add_event():
     if logged_in_cookie:
 
         title = request.forms.get('event_name')
+        description = request.forms.get('event_description')
+        all_day = request.forms.get('all_day')
         start_date = request.forms.get('start_date')
-        start_time = request.forms.get('start_time')
         end_date = request.forms.get('end_date')
-        end_time = request.forms.get('end_time')
 
-        # open Json FIle
-        with open('static/json/events.json', 'r') as file:
-            events = json.load(file)['events']
+        if all_day:
+            # open Json FIle
+            with open('static/json/events.json', 'r') as file:
+                events = json.load(file)['events']
 
-        # This creates incroment for id:s
-        for event in events:
-            max_id = int(event.get('id', 0))
-            id_for_new_event = max_id + 1
+            # This creates incroment for id:s
+            for event in events:
+                max_id = int(event.get('id', 0))
+                id_for_new_event = max_id + 1
 
-        add_event = {
-            "id": str(id_for_new_event),
-            "user_id": str(logged_in_cookie),
-            "title": title,
-            "start": f"{start_date}T{start_time}",
-            "end": f"{end_date}T{end_time}"
-        }
+            add_event = {
+                "id": str(id_for_new_event),
+                "user_id": str(logged_in_cookie),
+                "title": title,
+                "description": description,
+                "start": start_date,
+                "end": end_date
+            }
 
-        events.append(add_event)
+            events.append(add_event)
 
-        # Write updated events back to the JSON file
-        with open('static/json/events.json', 'w') as file:
-            json.dump({"events": events}, file, indent=4)
+            # Write updated events back to the JSON file
+            with open('static/json/events.json', 'w') as file:
+                json.dump({"events": events}, file, indent=4)
 
-        return redirect('/calendar') 
+            return redirect('/calendar')
+        else:
+            start_time = request.forms.get('start_time')
+            end_time = request.forms.get('end_time')
+
+            # open Json FIle
+            with open('static/json/events.json', 'r') as file:
+                events = json.load(file)['events']
+
+            # This creates incroment for id:s
+            for event in events:
+                max_id = int(event.get('id', 0))
+                id_for_new_event = max_id + 1
+
+            add_event = {
+                "id": str(id_for_new_event),
+                "user_id": str(logged_in_cookie),
+                "title": title,
+                "description": description,
+                "start": f"{start_date}T{start_time}",
+                "end": f"{end_date}T{end_time}"
+            }
+
+            events.append(add_event)
+
+            # Write updated events back to the JSON file
+            with open('static/json/events.json', 'w') as file:
+                json.dump({"events": events}, file, indent=4)
+
+            return redirect('/calendar') 
 
     else:
         return redirect('/')
-
-@route('/add_event', method=['GET', 'POST'])
-def add_event():
+    
+@route('/edit/event', method=['GET', 'POST'])
+def edit_event():
     '''
-        This function adds events to the events.json file
+        This function edits events
     '''
     logged_in_cookie = request.get_cookie('loggedIn')
     if logged_in_cookie:
 
-        title = request.forms.get('event_name')
-        start_date = request.forms.get('start_date')
-        start_time = request.forms.get('start_time')
-        end_date = request.forms.get('end_date')
-        end_time = request.forms.get('end_time')
+        id = request.forms.get('edit_event_id')
+        title = request.forms.get('title')
+        description = request.forms.get('description')
+        start_date = request.forms.get('start_date_edit')
+        start_time = request.forms.get('start_time_edit')
+        end_date = request.forms.get('end_date_edit')
+        end_time = request.forms.get('end_time_edit')
 
         # open Json FIle
         with open('static/json/events.json', 'r') as file:
@@ -438,18 +481,11 @@ def add_event():
 
         # This creates incroment for id:s
         for event in events:
-            max_id = int(event.get('id', 0))
-            id_for_new_event = max_id + 1
-
-        add_event = {
-            "id": str(id_for_new_event),
-            "user_id": str(logged_in_cookie),
-            "title": title,
-            "start": f"{start_date}T{start_time}",
-            "end": f"{end_date}T{end_time}"
-        }
-
-        events.append(add_event)
+            if event["id"] == id:
+                event["title"] = title
+                event["description"] = description
+                event["start"] = f"{start_date}T{start_time}"
+                event["end"] = f"{end_date}T{end_time}"
 
         # Write updated events back to the JSON file
         with open('static/json/events.json', 'w') as file:
@@ -504,6 +540,8 @@ def get_events():
         return json.dumps(filtered_events)
     else:
         return redirect('/')
+    
+# Calendar End
     
 @route('/progress_table')
 def progress_table():
